requirements_cpu.txt
filelock==3.18.0
fsspec==2025.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
setuptools==80.9.0
sympy==1.14.0
tqdm==4.67.1
typing_extensions==4.14.1
pyntcloud==0.3.1
plyfile==1.0.1
torchvision==0.22.1
torchac==0.9.3
scikit-learn==1.4.0
pandas==1.5.3
# Point cloud + visual tools
trimesh
pyyaml
seaborn
